djangoapp_api_clone/djangoapp/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djangorestframework

urllib3==1.26.7
beautifulsoup4
cfscrape

django-debug-toolbar

aiocfscrape
aiohttp >=3.1.3, <4a
js2py